create types and dont pass abortController back and forth
