+++ packages/app/src/cli/services/function/ui/components/Replay/types.ts
@@ -1,0 +1,17 @@
+export interface FunctionRunFromRunner {
+  type: 'functionRun'
+  input: string
+  output: string
+  logs: string
+  name: string
+  size: number
+  memory_usage: number
+  instructions: number
+}
+
+export interface SystemMessage {
+  type: 'systemMessage'
+  message: string
+}
+
+export type ReplayLog = FunctionRunFromRunner | SystemMessage
